+++ examples/example-auth/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/chapter-three/next-drupal/compare/example-auth@1.3.1...example-auth@1.3.2) (2022-09-07)
+
+**Note:** Version bump only for package example-auth
+
+
+
+
+
 ## [1.3.1](https://github.com/chapter-three/next-drupal/compare/example-auth@1.3.0...example-auth@1.3.1) (2022-07-29)
 
 **Note:** Version bump only for package example-auth