+++ examples/example-query/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/chapter-three/next-drupal/compare/example-query@0.3.0...example-query@0.3.1) (2022-09-07)
+
+
+### Bug Fixes
+
+* **example-query:** use context for fetching content ([99bf2b6](https://github.com/chapter-three/next-drupal/commit/99bf2b6afaf55d5955d03971d186b7da42ce8f35))
+
+
+
+
+
+# [0.3.0](https://github.com/chapter-three/next-drupal/compare/example-query@0.2.1...example-query@0.3.0) (2022-08-11)
+
+
+### Features
+
+* **example-query:** add example query with opts ([73b671e](https://github.com/chapter-three/next-drupal/commit/73b671eea6afae1328caf61379f4ae60529b40f8))
+
+
+
+
+
+## [0.2.1](https://github.com/chapter-three/next-drupal/compare/example-query@0.2.0...example-query@0.2.1) (2022-08-09)
+
+**Note:** Version bump only for package example-query
+
+
+
+
+
+# [0.2.0](https://github.com/chapter-three/next-drupal/compare/example-query@0.1.0...example-query@0.2.0) (2022-08-09)
+
+
+### Features
+
+* **example-query:** add example for nested queries ([9e28c2a](https://github.com/chapter-three/next-drupal/commit/9e28c2aac149505d84c7d24617c51dc2af052b0c))
+* **example-query:** update example to use formatters ([7e29ecb](https://github.com/chapter-three/next-drupal/commit/7e29ecb1e43171f3f8628e8126c55c73c96966be))
+
+
+
+
+
 # 0.1.0 (2022-07-29)